refactor: decompose large CLIcore source files into smaller units\n\nSplit 4 large CLIcore source files into smaller, descriptively\nnamed files to improve readability and maintainability:\n\nCLIcore_UI.c (4455 -> 3028 lines):\n  - CLIcore_UI_completion.c: readline/tab-completion (853 lines)\n  - CLIcore_UI_highlight.c: syntax coloring (151 lines)\n  - CLIcore_UI_hintarea.c: ghost text + hint area (721 lines)\n\nCLIcore_help.c (2664 -> 1387 lines):\n  - CLIcore_help_topics.c: help topic pages (750 lines)\n  - CLIcore_help_tui.c: interactive fuzzy help (609 lines)\n\nCLIcore_script.c (4022 -> 3115 lines):\n  - CLIcore_script_flow.c: if/while/for/select blocks (671 lines)\n  - CLIcore_script_case.c: case/esac + func define (309 lines)\n\nCLIcore_checkargs.c (1640 -> 921 lines):\n  - CLIcore_checkargs_fps.c: FPS param bridging (767 lines)\n\nUpdated CMakeLists.txt, CLIcore_UI.h, and CLIcore_script.h\nwith new source files and cross-file declarations."
