+++ src/cli/CLIcore/CLIcore/CLIcore_UI.c
@@ -47,675 +47,8 @@
 extern void yy_scan_string(const char *);
 extern int  yylex_destroy(void);
 
-#ifdef USE_READLINE
-/**
- * @brief Readline callback
- *
- **/
-/**
- * @brief Custom getc wrapper for readline
- *
- * text (printed after the cursor by print_ghost)
- * BEFORE readline processes the newline.
- */
-
-/**
- * Number of ghost chars rendered on current line.
- * Set by print_ghost(), read by cli_accept_line().
- */
-int ghost_chars_on_line = 0;
-
-/**
- * @brief Custom accept-line handler for readline
- *
- * Bound to Enter key. Overwrites ghost suggestion
- * text with spaces before accepting the line, so
- * the terminal scrollback entry is clean.
- */
-int cli_accept_line(
-    int count,
-    int key
-)
-{
-    if(ghost_chars_on_line > 0)
-    {
-        int n = ghost_chars_on_line;
-        for(int i = 0; i < n; i++)
-        {
-            putchar(' ');
-        }
-        for(int i = 0; i < n; i++)
-        {
-            putchar('\b');
-        }
-        fflush(stdout);
-        ghost_chars_on_line = 0;
-    }
-
-    return rl_newline(count, key);
-}
-
-void rl_cb_linehandler(char *linein)
-{
-    if(NULL == linein)
-    {
-        data.CLIloopON = 0;
-        return;
-    }
-
-    data.CLIexecuteCMDready = 1;
-
-    // copy input into data.CLIcmdline
-    strcpy(data.CLIcmdline, linein);
-
-    /* Handle backslash line continuation:
-     * temporarily switch to blocking readline
-     * to read additional lines */
-    {
-        size_t len = strlen(data.CLIcmdline);
-        while(len > 0
-                && data.CLIcmdline[len - 1]
-                == '\\')
-        {
-            data.CLIcmdline[len - 1] = ' ';
-            /* Remove callback handler to avoid
-             * interference, use direct readline
-             * for continuation */
-            rl_callback_handler_remove();
-            char *cont = readline("> ");
-            /* Re-install with dummy prompt;
-             * the main loop will re-install
-             * with the proper prompt after this
-             * handler returns */
-            rl_callback_handler_install(
-                "",
-                (rl_vcpfunc_t *)
-                &rl_cb_linehandler);
-            if(cont == NULL)
-            {
-                break;
-            }
-            int avail =
-                STRINGMAXLEN_CLICMDLINE
-                - (int) strlen(data.CLIcmdline)
-                - 1;
-            if(avail > 0)
-            {
-                strncat(data.CLIcmdline, cont,
-                        (size_t) avail);
-            }
-            free(cont);
-            len = strlen(data.CLIcmdline);
-        }
-    }
-
-    CLI_execute_line();
-
-    free(linein);
-}
-#endif
-
-errno_t runCLI_prompt(char *promptstring, char *prompt)
-{
-    // Try to get PS1 from CLI vars or environment
-    const char *ps1_val = cli_var_lookup("PS1");
-    if(ps1_val == NULL)
-    {
-        ps1_val = getenv("PS1");
-    }
-
-    if(ps1_val != NULL && strlen(ps1_val) > 0)
-    {
-        char expanded_ps1[FPS_DIR_STRLENMAX];
-        strncpy(expanded_ps1, ps1_val, FPS_DIR_STRLENMAX - 1);
-        expanded_ps1[FPS_DIR_STRLENMAX - 1] = '\0';
-        cli_expand_env(expanded_ps1, FPS_DIR_STRLENMAX);
-        strncpy(prompt, expanded_ps1, FPS_DIR_STRLENMAX - 1);
-        prompt[FPS_DIR_STRLENMAX - 1] = '\0';
-        return RETURN_SUCCESS;
-    }
-
-    if(strlen(promptstring) > 0)
-    {
-        if(data.processnameflag == 0)
-        {
-            snprintf(prompt, FPS_DIR_STRLENMAX, COLORHBOLDCYAN "%s > " RL_COLORRESET,
-                     promptstring);
-        }
-        else
-        {
-            snprintf(prompt,
-                     FPS_DIR_STRLENMAX,
-                     COLORHBOLDCYAN "%s-%s > " RL_COLORRESET,
-                     promptstring,
-                     data.processname);
-        }
-    }
-    else
-    {
-        snprintf(prompt, FPS_DIR_STRLENMAX, COLORHBOLDCYAN "%s > " RL_COLORRESET,
-                 data.processname);
-    }
-
-    return RETURN_SUCCESS;
-}
 
 
-void *xmalloc(int size)
-{
-    void *buf;
-
-    buf = malloc(size);
-    if(!buf)
-    {
-        fprintf(stderr, COLORRED "Error: Out of memory. Exiting.'n" COLORRESET);
-        exit(1);
-    }
-
-    return buf;
-}
-
-char *dupstr(char *s)
-{
-    char *r;
-
-    r = (char *) xmalloc((strlen(s) + 1));
-    strcpy(r, s);
-    return (r);
-}
-
-#ifdef USE_READLINE
-int levenshtein_distance(const char *s1, const char *s2)
-{
-    unsigned int len1 = strlen(s1);
-    unsigned int len2 = strlen(s2);
-    unsigned int *d = (unsigned int *)xmalloc((len1 + 1) * (len2 + 1) * sizeof(unsigned int));
-
-    for(unsigned int i = 0; i <= len1; i++) d[i * (len2 + 1)] = i;
-    for(unsigned int j = 0; j <= len2; j++) d[j] = j;
-
-    for(unsigned int i = 1; i <= len1; i++) {
-        for(unsigned int j = 1; j <= len2; j++) {
-            unsigned int cost = (s1[i - 1] == s2[j - 1]) ? 0 : 1;
-            unsigned int min1 = d[(i - 1) * (len2 + 1) + j] + 1;
-            unsigned int min2 = d[i * (len2 + 1) + j - 1] + 1;
-            unsigned int min3 = d[(i - 1) * (len2 + 1) + j - 1] + cost;
-            unsigned int m = (min1 < min2) ? min1 : min2;
-            d[i * (len2 + 1) + j] = (m < min3) ? m : min3;
-        }
-    }
-    int dist = d[len1 * (len2 + 1) + len2];
-    free(d);
-    return dist;
-}
-
-/**
- * @brief State for fuzzy fallback pass in generator
- *
- * After a normal prefix-match pass, if nothing matched
- * and fuzzy is enabled, we restart with substring match.
- */
-int generator_fuzzy_pass = 0;
-
-char *CLI_generator(const char *text, int state)
-{
-    static unsigned int list_index;
-    static unsigned int len;
-    char               *name;
-
-    if(!state)
-    {
-        list_index  = 0;
-        len         = strlen(text);
-        generator_fuzzy_pass = 0;
-    }
-
-retry_fuzzy:
-
-    if(data.CLImatchMode == CLICOMPLETIONMODE_COMMANDS)
-    {
-        while(list_index < data.NBcmd)
-        {
-            name = data.cmd[list_index].key;
-            list_index++;
-            if(generator_fuzzy_pass == 0)
-            {
-                if(strncmp(name, text, len) == 0)
-                {
-                    return (dupstr(name));
-                }
-            }
-            else
-            {
-                /* Fuzzy: substring match */
-                if(strstr(name, text) != NULL)
-                {
-                    return (dupstr(name));
-                }
-            }
-        }
-    }
-
-    if(data.CLImatchMode == CLICOMPLETIONMODE_IMAGES)
-    {
-        static DIR *img_dirp = NULL;
-
-        if(!state)
-        {
-            if(img_dirp != NULL)
-            {
-                closedir(img_dirp);
-                img_dirp = NULL;
-            }
-            img_dirp = opendir(dcshmdir);
-        }
-
-        if(img_dirp != NULL)
-        {
-            struct dirent *ent;
-            while((ent = readdir(img_dirp)) != NULL)
-            {
-                char *ext = strstr(ent->d_name, ".im.shm");
-                if(ext != NULL && strcmp(ext, ".im.shm") == 0)
-                {
-                    char imgname[256];
-                    int namelen = ext - ent->d_name;
-                    if(namelen > 255)
-                    {
-                        namelen = 255;
-                    }
-                    strncpy(imgname, ent->d_name, namelen);
-                    imgname[namelen] = '\0';
-
-                    if(generator_fuzzy_pass == 0)
-                    {
-                        if(strncmp(imgname, text, len) == 0)
-                        {
-                            return (dupstr(imgname));
-                        }
-                    }
-                    else
-                    {
-                        if(strstr(imgname, text) != NULL)
-                        {
-                            return (dupstr(imgname));
-                        }
-                    }
-                }
-            }
-            closedir(img_dirp);
-            img_dirp = NULL;
-        }
-    }
-
-    if(data.CLImatchMode == CLICOMPLETIONMODE_CMDARGS)
-    {
-        while((int) list_index <
-                data.cmd[data.cmdindex].nbarg)
-        {
-            name = data.cmd[data.cmdindex]
-                       .argdata[list_index]
-                       .fpstag;
-            list_index++;
-            if(strncmp(name, text, len) == 0)
-            {
-                return (dupstr(name));
-            }
-        }
-    }
-
-    if(data.CLImatchMode == CLICOMPLETIONMODE_FILES)
-    {
-        /* Filesystem path completion.
-         * Split text into directory + prefix,
-         * enumerate with opendir/readdir. */
-        static DIR *dirp = NULL;
-        static char dirpart[512];
-        static char prefix[256];
-        static unsigned int preflen;
-
-        if(!state)
-        {
-            if(dirp != NULL)
-            {
-                closedir(dirp);
-                dirp = NULL;
-            }
-
-            /* Split text into dir + filename prefix */
-            const char *slash =
-                strrchr(text, '/');
-            if(slash != NULL)
-            {
-                int dlen = (int)(slash - text) + 1;
-                if(dlen > (int) sizeof(dirpart) - 1)
-                {
-                    dlen = (int) sizeof(dirpart) - 1;
-                }
-                memcpy(dirpart, text, dlen);
-                dirpart[dlen] = '\0';
-                strncpy(prefix, slash + 1,
-                        sizeof(prefix) - 1);
-                prefix[sizeof(prefix) - 1] = '\0';
-            }
-            else
-            {
-                strcpy(dirpart, ".");
-                strncpy(prefix, text,
-                        sizeof(prefix) - 1);
-                prefix[sizeof(prefix) - 1] = '\0';
-            }
-            preflen = strlen(prefix);
-
-            dirp = opendir(dirpart);
-        }
-
-        if(dirp != NULL)
-        {
-            struct dirent *ent;
-            while((ent = readdir(dirp)) != NULL)
-            {
-                /* Skip . and .. */
-                if(strcmp(ent->d_name, ".") == 0
-                    || strcmp(ent->d_name,
-                             "..") == 0)
-                {
-                    continue;
-                }
-
-                if(strncmp(ent->d_name, prefix,
-                           preflen) == 0)
-                {
-                    /* Build full path for
-                     * stat check */
-                    char fullpath[1024];
-                    snprintf(fullpath,
-                             sizeof(fullpath),
-                             "%s/%s",
-                             dirpart,
-                             ent->d_name);
-
-                    /* Build result: dirpart
-                     * prefix + entry name */
-                    char result[1024];
-                    if(strcmp(dirpart, ".") == 0)
-                    {
-                        snprintf(result,
-                                 sizeof(result),
-                                 "%s",
-                                 ent->d_name);
-                    }
-                    else
-                    {
-                        snprintf(result,
-                                 sizeof(result),
-                                 "%s%s",
-                                 dirpart,
-                                 ent->d_name);
-                    }
-
-                    /* Append / for directories */
-                    struct stat st;
-                    if(stat(fullpath, &st) == 0
-                        && S_ISDIR(st.st_mode))
-                    {
-                        strncat(result, "/",
-                                sizeof(result)
-                                - strlen(result)
-                                - 1);
-                    }
-
-                    return dupstr(result);
-                }
-            }
-            closedir(dirp);
-            dirp = NULL;
-        }
-    }
-
-    if(data.CLImatchMode == CLICOMPLETIONMODE_FPSPARAMS)
-    {
-        /*
-         * FPS name completion.
-         * Scan dcshmdir for fps.*.datadir entries,
-         * strip "fps." prefix and ".datadir" suffix.
-         */
-        static DIR *fps_dirp = NULL;
-
-        if(!state)
-        {
-            if(fps_dirp != NULL)
-            {
-                closedir(fps_dirp);
-                fps_dirp = NULL;
-            }
-            fps_dirp = opendir(dcshmdir);
-        }
-
-        if(fps_dirp != NULL)
-        {
-            struct dirent *ent;
-            while((ent = readdir(fps_dirp)) != NULL)
-            {
-                /* Only fps.*.datadir entries */
-                if(strncmp(ent->d_name, "fps.", 4) != 0)
-                {
-                    continue;
-                }
-                char *ext = strstr(ent->d_name, ".datadir");
-                if(ext != NULL && strcmp(ext, ".datadir") == 0)
-                {
-                    char fpsname[256];
-                    int namelen = ext - (ent->d_name + 4);
-                    if(namelen > 255)
-                    {
-                        namelen = 255;
-                    }
-                    strncpy(fpsname, ent->d_name + 4, namelen);
-                    fpsname[namelen] = '\0';
-
-                    if(generator_fuzzy_pass == 0)
-                    {
-                        if(strncmp(fpsname, text, len) == 0)
-                        {
-                            return dupstr(fpsname);
-                        }
-                    }
-                    else
-                    {
-                        if(strstr(fpsname, text) != NULL)
-                        {
-                            return dupstr(fpsname);
-                        }
-                    }
-                }
-            }
-            closedir(fps_dirp);
-            fps_dirp = NULL;
-        }
-    }
-
-    /* Fuzzy fallback: if prefix pass found nothing,
-     * restart with substring matching */
-    if(generator_fuzzy_pass == 0 &&
-            data.autocomplete_fuzzy)
-    {
-        generator_fuzzy_pass = 1;
-        list_index  = 0;
-        goto retry_fuzzy;
-    }
-
-    return ((char *) NULL);
-}
-
-/** @brief readline custom completion
- *
- * Invoked when pressing TAB
- */
-
-char **
-CLI_completion(const char *text, int start, int __attribute__((unused)) end)
-{
-    char **matches;
-
-    //printf("[%d | %s | %s]", start, rl_line_buffer, text);
-    //rl_message("\n[%d %s]\n", start, rl_line_buffer);
-    //rl_redisplay();
-    //rl_forced_update_display();
-
-    matches = (char **) NULL;
-
-    if((start == 0) || (strncmp(rl_line_buffer, "cmd?", strlen("cmd?")) == 0))
-    {
-        // if first word, or second argument to cmd?, match string with commands
-        data.CLImatchMode = CLICOMPLETIONMODE_COMMANDS;
-    }
-    else
-    {
-        // test if first word is a command
-        char  str[200];
-        char *firstword;
-        firstword = strcpy(str, rl_line_buffer);
-        strtok(str, " ");
-        int      cmdimatch = -1;
-        uint32_t cmdi      = 0;
-        while((cmdimatch == -1) && (cmdi < data.NBcmd))
-        {
-            if(strcmp(firstword, data.cmd[cmdi].key) == 0)
-            {
-                cmdimatch = cmdi;
-                //printf("COMMAND MATCH %s\n", data.cmd[cmdi].key);
-                data.cmdindex = cmdi;
-            }
-            cmdi++;
-        }
-
-        if((cmdimatch != -1) && (text[0] == '.'))
-        {
-            data.CLImatchMode = CLICOMPLETIONMODE_CMDARGS;
-        }
-        else if(cmdimatch != -1)
-        {
-            /* Count which CLI argument position
-             * the cursor is at */
-            int argpos = 0;
-            {
-                const char *p = rl_line_buffer;
-                /* Skip command word */
-                while(*p && *p != ' ')
-                {
-                    p++;
-                }
-                int in_word = 0;
-                while(*p)
-                {
-                    if(*p != ' ')
-                    {
-                        if(!in_word)
-                        {
-                            argpos++;
-                            in_word = 1;
-                        }
-                    }
-                    else
-                    {
-                        in_word = 0;
-                    }
-                    p++;
-                }
-                /* If trailing space, next arg */
-                if(rl_end > 0
-                    && rl_line_buffer[
-                        rl_end - 1] == ' ')
-                {
-                    /* argpos already correct */
-                }
-                else if(argpos > 0)
-                {
-                    argpos--;
-                }
-            }
-
-            /* Check argument type at this
-             * position (CLI-visible args) */
-            int cli_ai = 0;
-            int matched_file = 0;
-            for(int ai = 0;
-                ai < data.cmd[cmdimatch].nbparam;
-                ai++)
-            {
-                if(data.cmd[cmdimatch]
-                    .argdata[ai].fpflag
-                    & FPFLAG_PRIMARY_CLI_INPUT)
-                {
-                    if(cli_ai == argpos)
-                    {
-                        uint64_t atype =
-                            data.cmd[cmdimatch]
-                            .argdata[ai].type;
-                        if(atype == CLIARG_FILENAME
-                            || atype
-                            == CLIARG_FITSFILENAME)
-                        {
-                            matched_file = 1;
-                        }
-                        if(atype == CLIARG_FPSNAME)
-                        {
-                            data.CLImatchMode =
-                                CLICOMPLETIONMODE_FPSPARAMS;
-                        }
-                        break;
-                    }
-                    cli_ai++;
-                }
-            }
-
-            if(matched_file)
-            {
-                data.CLImatchMode =
-                    CLICOMPLETIONMODE_FILES;
-                /* Don't append space after
-                 * directory names */
-                rl_completion_append_character
-                    = '\0';
-            }
-            else if(data.CLImatchMode
-                    != CLICOMPLETIONMODE_FPSPARAMS)
-            {
-                if(strcmp(data.cmd[cmdimatch].key, "fparam") == 0 ||
-                   strcmp(data.cmd[cmdimatch].key, "fpsCTRL") == 0 ||
-                   strcmp(data.cmd[cmdimatch].key, "fpsload") == 0 ||
-                   strcmp(data.cmd[cmdimatch].key, "dpsingle") == 0)
-                {
-                    data.CLImatchMode = CLICOMPLETIONMODE_FPSPARAMS;
-                }
-                else
-                {
-                    data.CLImatchMode = CLICOMPLETIONMODE_IMAGES;
-                }
-            }
-        }
-        else
-        {
-            // match string with images
-            data.CLImatchMode = CLICOMPLETIONMODE_IMAGES;
-        }
-    }
-
-    matches = rl_completion_matches((char *) text, &CLI_generator);
-
-    /* Reset append char to default space */
-    if(data.CLImatchMode != CLICOMPLETIONMODE_FILES)
-    {
-        rl_completion_append_character = ' ';
-    }
-
-    //    else
-    //  rl_bind_key('\t',rl_abort);
-
-    return (matches);
-}
-#endif
 
 errno_t write_tracedebugfile()
 {
@@ -777,108 +110,6 @@ errno_t write_tracedebugfile()
 
     return RETURN_SUCCESS;
 }
-
-int cli_is_command(const char *word)
-{
-    for(uint32_t i = 0; i < data.NBcmd; i++)
-    {
-        if(strcmp(data.cmd[i].key, word) == 0)
-        {
-            return 1;
-        }
-    }
-    return 0;
-}
-
-void cli_highlight_redisplay(void)
-{
-    if(!data.syntax_highlight)
-    {
-        rl_redisplay();
-        return;
-    }
-
-    /*
-     * Let readline draw normally first so its
-     * internal cursor state stays consistent.
-     * Then overwrite just the first word in color.
-     */
-    rl_redisplay();
-
-    /* Find the first word boundaries */
-    int ws = 0;
-    while(rl_line_buffer[ws] == ' '
-            || rl_line_buffer[ws] == '\t')
-    {
-        ws++;
-    }
-    int we = ws;
-    while(rl_line_buffer[we] != '\0'
-            && rl_line_buffer[we] != ' '
-            && rl_line_buffer[we] != '\t')
-    {
-        we++;
-    }
-    if(we == ws)
-    {
-        fflush(stdout);
-        return;
-    }
-
-    /* Extract first word */
-    char firstword[200];
-    int fwlen = we - ws;
-    if(fwlen > 199)
-    {
-        fwlen = 199;
-    }
-    memcpy(firstword, rl_line_buffer + ws,
-           (size_t) fwlen);
-    firstword[fwlen] = '\0';
-
-    /* Pick color */
-    const char *col;
-    if(cli_is_command(firstword))
-    {
-        col = "\033[32m"; /* green */
-    }
-    else
-    {
-        col = "\033[31m"; /* red */
-    }
-
-    /*
-     * After rl_redisplay(), cursor is at rl_point.
-     * Move back to the first word, overwrite with
-     * color, then restore cursor position.
-     * Use cursor-relative movement (not absolute
-     * column) to avoid prompt-width errors from
-     * invisible escape sequences in the prompt.
-     */
-    fprintf(rl_outstream, "\033[s");  /* ANSI save */
-    {
-        int back = rl_point - ws;
-        if(back > 0)
-        {
-            fprintf(rl_outstream,
-                    "\033[%dD", back);
-        }
-    }
-    fprintf(rl_outstream, "%s%s\033[0m",
-            col, firstword);
-    fprintf(rl_outstream, "\033[u");  /* ANSI restore */
-    fflush(rl_outstream);
-}
-
-
-/*
- * ============================================================
- *  Command Bookmarks
- * ============================================================
- */
-
-
-
 
 
 
@@ -3794,662 +3025,4 @@ pipe_fallthrough:
     DEBUG_TRACE_FEXIT();
     return RETURN_SUCCESS;
 }
-
-#ifdef USE_READLINE
-
-/** @brief Stores the current inline suggestion suffix
- *
- * Set by CLI_redisplay when a suggestion is shown.
- * Consumed by accept_suggestion when Right Arrow is pressed.
- */
-char *pending_suggestion = NULL;
-int   pending_replace_len = 0;
-
-/**
- * @brief Accept the inline suggestion on Right Arrow
- *
- * If cursor is at end-of-line and a pending suggestion
- * exists, insert it. Otherwise, fall through to normal
- * cursor-right movement.
- */
-int accept_suggestion(
-    int count,
-    int key
-)
-{
-    (void) count;
-    (void) key;
-
-    if(pending_suggestion && rl_point == rl_end)
-    {
-        if(pending_replace_len > 0
-            && rl_end >= pending_replace_len)
-        {
-            int del_start =
-                rl_end - pending_replace_len;
-            rl_delete_text(del_start, rl_end);
-            rl_point = del_start;
-        }
-        rl_insert_text(pending_suggestion);
-        free(pending_suggestion);
-        pending_suggestion = NULL;
-        pending_replace_len = 0;
-        rl_redisplay();
-        return 0;
-    }
-
-    /* Not at EOL or no suggestion — normal right */
-    return rl_forward_char(1, key);
-}
-
-/**
- * @brief Store the suggestion suffix for Right Arrow
- */
-void set_pending_suggestion(
-    const char *text,
-    int replace_len
-)
-{
-    free(pending_suggestion);
-    pending_suggestion = NULL;
-    pending_replace_len = 0;
-    if(text && strlen(text) > 0)
-    {
-        pending_suggestion =
-            dupstr((char *) text);
-        pending_replace_len = replace_len;
-    }
-}
-
-/**
- * @brief Find the command index matching `firstword`
- *
- * Returns -1 if no match.
- */
-int find_command_match(const char *firstword)
-{
-    for(uint32_t cmdi = 0; cmdi < data.NBcmd; cmdi++)
-    {
-        if(strcmp(firstword, data.cmd[cmdi].key) == 0)
-        {
-            data.cmdindex = cmdi;
-            return (int) cmdi;
-        }
-    }
-    return -1;
-}
-
-/**
- * @brief Compute visible length of readline prompt
- *
- * Strips \001..\002 escape wrappers that readline
- * uses to mark non-printing characters.
- */
-int visible_prompt_len(void)
-{
-    const char *p = rl_display_prompt
-                        ? rl_display_prompt
-                        : "";
-    int   len = 0;
-    int   invisible = 0;
-
-    for(; *p; p++)
-    {
-        if(*p == '\001')
-        {
-            invisible = 1;
-        }
-        else if(*p == '\002')
-        {
-            invisible = 0;
-        }
-        else if(!invisible)
-        {
-            len++;
-        }
-    }
-    return len;
-}
-
-/**
- * @brief Get number of ghost chars we can print
- *
- * Returns max chars that can be printed after the
- * cursor without wrapping to the next terminal line.
- */
-int get_ghost_budget(void)
-{
-    struct winsize ws;
-    int cols = 80; /* fallback */
-
-    if(ioctl(STDOUT_FILENO, TIOCGWINSZ, &ws) >= 0
-            && ws.ws_col > 0)
-    {
-        cols = ws.ws_col;
-    }
-
-    int cursor_col =
-        (visible_prompt_len() + rl_point) % cols;
-
-    int budget = cols - cursor_col - 1;
-    if(budget < 0)
-    {
-        budget = 0;
-    }
-    return budget;
-}
-
-/**
- * @brief Print ghost text with truncation
- *
- * Prints up to budget visible chars from text
- * in the given ANSI style, then resets style.
- * Returns number of visible chars printed.
- */
-int print_ghost(
-    const char *style,
-    const char *text,
-    int budget
-)
-{
-    int tlen = (int) strlen(text);
-    int plen = tlen < budget ? tlen : budget;
-
-    if(plen <= 0)
-    {
-        return 0;
-    }
-
-    printf("%s%.*s\033[0m", style, plen, text);
-    ghost_chars_on_line = plen;
-    return plen;
-}
-
-/**
- * @brief State for the reserved hint area
- */
-int hint_area_active = 0;
-int cached_term_rows = 0;
-int cached_term_cols = 0;
-
-/**
- * @brief Set up scroll region reserving bottom line
- *
- * Confines normal terminal output to lines 1..(rows-1)
- * so the bottom line stays fixed for hints.
- */
-void CLI_setup_hint_area(void)
-{
-    struct winsize ws;
-    if(ioctl(STDOUT_FILENO, TIOCGWINSZ, &ws) < 0
-            || ws.ws_row <= 3)
-    {
-        hint_area_active = 0;
-        return;
-    }
-
-    cached_term_rows = ws.ws_row;
-    cached_term_cols = ws.ws_col;
-
-    /* Ensure there's a free line below the cursor.
-     * If the cursor is at the bottom of the screen, \n scrolls
-     * the screen up by 1. \033[1A moves it back to its relative
-     * position. This ensures the cursor is never at 'rows'
-     * (the reserved hint line) before we save its position. */
-    printf("\n\033[1A");
-
-    /* Save the current cursor position using ANSI */
-    printf("\033[s");
-
-    /* Set scroll region to rows 1..(rows-1)
-     * NOTE: DECSTBM moves cursor to home (1, 1) */
-    printf("\033[1;%dr", cached_term_rows - 1);
-
-    /* Clear the hint line (outside scroll region) */
-    printf("\033[%d;1H\033[2K", cached_term_rows);
-
-    /* Restore cursor to where it was using ANSI */
-    printf("\033[u");
-    fflush(stdout);
-
-    hint_area_active = 1;
-}
-
-/**
- * @brief Reset scroll region to full terminal
- *
- * Call this before exiting readline mode or
- * when the CLI session ends.
- */
-void CLI_cleanup_scroll_region(void)
-{
-    if(!hint_area_active)
-    {
-        return;
-    }
-
-    /* Save cursor position (ANSI — does NOT
-     * touch scroll margins) */
-    printf("\033[s");
-
-    /* Move to hint line and erase it BEFORE
-     * resetting scroll region, while the row
-     * is still addressable at its cached pos */
-    printf("\033[%d;1H\033[2K", cached_term_rows);
-
-    /* Reset scroll region to full terminal.
-     * This also moves cursor to (1,1). */
-    printf("\033[r");
-
-    /* Restore cursor to where it was before */
-    printf("\033[u");
-
-    fflush(stdout);
-
-    hint_area_active = 0;
-}
-
-/**
- * @brief Update the hint area with function prototype
- *
- * Paints the reserved bottom line with the command
- * syntax when a known command is being typed.
- */
-void update_hint_area(void)
-{
-    if(!hint_area_active || !data.autocomplete_arghint)
-    {
-        return;
-    }
-
-    /* Single ANSI save cursor for the whole
-     * operation (resize + hint painting) */
-    printf("\033[s");
-
-    /* Check for terminal resize */
-    {
-        struct winsize ws;
-        if(ioctl(STDOUT_FILENO, TIOCGWINSZ,
-                 &ws) >= 0 &&
-                ws.ws_row > 3 &&
-                (ws.ws_row != cached_term_rows ||
-                 ws.ws_col != cached_term_cols))
-        {
-            cached_term_rows = ws.ws_row;
-            cached_term_cols = ws.ws_col;
-            /* Re-set scroll region (cursor jumps
-             * to home, but we saved it above) */
-            printf("\033[1;%dr",
-                   cached_term_rows - 1);
-            /* Clear new hint line */
-            printf("\033[%d;1H\033[2K",
-                   cached_term_rows);
-        }
-    }
-
-    /* Move to hint line, clear it */
-    printf("\033[%d;1H\033[2K",
-           cached_term_rows);
-
-    /* Check if first word is a known command */
-    if(rl_line_buffer[0] != '\0')
-    {
-        char  buf[200];
-        char *saveptr_hint = NULL;
-        snprintf(buf, sizeof(buf), "%s",
-                 rl_line_buffer);
-        char *fw = strtok_r(
-                       buf, " ", &saveptr_hint);
-
-        if(fw != NULL)
-        {
-            int cmi = find_command_match(fw);
-            if(cmi >= 0)
-            {
-                /* Count argument words after cmd
-                 * to determine current arg */
-                int argidx = 0;
-                {
-                    const char *p =
-                        rl_line_buffer;
-                    /* Skip command word */
-                    while(*p && *p != ' ')
-                    {
-                        p++;
-                    }
-                    /* Count argument words */
-                    int wcount = 0;
-                    int in_word = 0;
-                    while(*p)
-                    {
-                        if(*p != ' ')
-                        {
-                            if(!in_word)
-                            {
-                                wcount++;
-                                in_word = 1;
-                            }
-                        }
-                        else
-                        {
-                            in_word = 0;
-                        }
-                        p++;
-                    }
-                    if(rl_end > 0 &&
-                            rl_line_buffer[
-                                rl_end - 1]
-                            == ' ')
-                    {
-                        /* Trailing space: about
-                         * to type next arg */
-                        argidx = wcount;
-                    }
-                    else
-                    {
-                        /* Mid-word: typing this
-                         * arg (0-indexed) */
-                        argidx = wcount > 0
-                                 ? wcount - 1 : 0;
-                    }
-                }
-
-                /* Print syntax with <> delimited
-                 * tokens, highlighting active */
-                const char *syn =
-                    data.cmd[cmi].syntax;
-                int col = 0;
-                int tidx = 0;
-                const char *p = syn;
-
-                while(*p &&
-                        col < cached_term_cols - 2)
-                {
-                    /* Skip whitespace between
-                     * argument tokens */
-                    if(*p == ' ')
-                    {
-                        printf(" ");
-                        col++;
-                        p++;
-                        continue;
-                    }
-
-                    /* Find extent of this token */
-                    const char *tstart = p;
-                    if(*p == '<')
-                    {
-                        /* Scan to matching '>' */
-                        while(*p && *p != '>')
-                        {
-                            p++;
-                        }
-                        if(*p == '>')
-                        {
-                            p++;
-                        }
-                    }
-                    else
-                    {
-                        /* Non-<> word */
-                        while(*p && *p != ' '
-                                && *p != '<')
-                        {
-                            p++;
-                        }
-                    }
-                    int tlen = (int)(p - tstart);
-                    int avail =
-                        cached_term_cols - 1 - col;
-                    int plen = tlen < avail
-                               ? tlen : avail;
-
-                    if(*tstart == '<' &&
-                            tidx == argidx)
-                    {
-                        printf("\033[1;97m"
-                               "%.*s"
-                               "\033[0m",
-                               plen, tstart);
-                    }
-                    else
-                    {
-                        printf("\033[2m"
-                               "%.*s"
-                               "\033[0m",
-                               plen, tstart);
-                    }
-                    col += plen;
-                    if(*tstart == '<')
-                    {
-                        tidx++;
-                    }
-                }
-            }
-        }
-    }
-
-    /* ANSI restore cursor */
-    printf("\033[u");
-    fflush(stdout);
-}
-
-void CLI_redisplay(void)
-{
-    /* Default or syntax-highlighted redisplay */
-    rl_redisplay_function = NULL;
-    if(data.syntax_highlight
-            && rl_line_buffer[0] != '\0')
-    {
-        cli_highlight_redisplay();
-    }
-    else
-    {
-        rl_redisplay();
-        fflush(stdout);
-    }
-    rl_redisplay_function = CLI_redisplay;
-
-    /* Clear any stale suggestion */
-    set_pending_suggestion(NULL, 0);
-
-    if(data.autocomplete == 0)
-    {
-        return;
-    }
-
-    if(rl_line_buffer[0] == '\0')
-    {
-        update_hint_area();
-        return;
-    }
-
-
-    if(rl_point != rl_end)
-    {
-        /* Update hint even when not at EOL */
-        update_hint_area();
-        return;
-    }
-
-    int budget = get_ghost_budget();
-    if(budget <= 0)
-    {
-        update_hint_area();
-        return;
-    }
-
-    int total_ghost = 0;
-
-    /* ===== History-based suggestion ===== */
-    if(data.autocomplete_history)
-    {
-        HIST_ENTRY **hist = history_list();
-        if(hist)
-        {
-            int hlen = history_length;
-            for(int i = hlen - 1; i >= 0; i--)
-            {
-                if(strncmp(hist[i]->line,
-                           rl_line_buffer,
-                           rl_end) == 0 &&
-                        (int) strlen(
-                            hist[i]->line) >
-                        rl_end)
-                {
-                    const char *suffix =
-                        hist[i]->line + rl_end;
-                    int n = print_ghost(
-                                "\033[38;5;245m",
-                                suffix,
-                                budget);
-                    if(n > 0)
-                    {
-                        printf("\033[K");
-                        printf("\033[%dD", n);
-                        fflush(stdout);
-                        set_pending_suggestion(
-                            suffix, 0);
-                    }
-                    update_hint_area();
-                    return;
-                }
-            }
-        }
-    }
-
-    /* ===== Generator-based suggestion ===== */
-
-    /* Find current word start */
-    int start = 0;
-    for(int i = rl_point - 1; i >= 0; i--)
-    {
-        if(rl_line_buffer[i] == ' ')
-        {
-            start = i + 1;
-            break;
-        }
-    }
-
-    char *text = rl_line_buffer + start;
-
-    /* Determine matching mode */
-    if((start == 0) ||
-            (strncmp(rl_line_buffer, "cmd?",
-                     strlen("cmd?")) == 0))
-    {
-        data.CLImatchMode =
-            CLICOMPLETIONMODE_COMMANDS;
-    }
-    else
-    {
-        char  str[200];
-        char *saveptr_comp = NULL;
-        snprintf(str, 200, "%s", rl_line_buffer);
-        char *firstword = strtok_r(
-                str, " ", &saveptr_comp);
-
-        int cmdimatch = -1;
-        if(firstword != NULL)
-        {
-            cmdimatch =
-                find_command_match(firstword);
-        }
-
-        /* If command has no <> argument tokens,
-         * don't suggest arguments */
-        if(cmdimatch >= 0)
-        {
-            const char *syn =
-                data.cmd[cmdimatch].syntax;
-            if(syn == NULL ||
-                    strchr(syn, '<') == NULL)
-            {
-                update_hint_area();
-                return;
-            }
-        }
-
-        if((cmdimatch != -1) && (text[0] == '.'))
-        {
-            data.CLImatchMode =
-                CLICOMPLETIONMODE_CMDARGS;
-        }
-        else
-        {
-            data.CLImatchMode =
-                CLICOMPLETIONMODE_IMAGES;
-        }
-    }
-
-    /* Get best match */
-    char *match = CLI_generator(text, 0);
-
-    if(match)
-    {
-        if(strncmp(match, text, strlen(text)) == 0)
-        {
-            char *suffix = match + strlen(text);
-            int n = print_ghost(
-                        "\033[38;5;245m",
-                        suffix,
-                        budget);
-            if(n > 0)
-            {
-                total_ghost += n;
-                set_pending_suggestion(
-                    suffix, 0);
-            }
-        }
-        else if(data.autocomplete_fuzzy)
-        {
-            char fzbuf[256];
-            snprintf(fzbuf, sizeof(fzbuf),
-                     " [%s]", match);
-            int n = print_ghost(
-                        "\033[38;5;245m",
-                        fzbuf,
-                        budget);
-            total_ghost += n;
-            set_pending_suggestion(
-                match,
-                (int) strlen(text));
-        }
-        free(match);
-    }
-
-    /* Erase rest of line + move cursor back */
-    if(total_ghost > 0)
-    {
-        printf("\033[K");
-        printf("\033[%dD", total_ghost);
-        fflush(stdout);
-    }
-
-    update_hint_area();
-}
-
-void CLI_configure_readline()
-{
-    rl_redisplay_function = CLI_redisplay;
-
-    if(data.autocomplete_history)
-    {
-        read_history(CLI_history_file());
-    }
-
-    /* Bind Right Arrow to accept suggestion
-     * when at end-of-line */
-    rl_bind_keyseq("\\e[C", accept_suggestion);
-
-    /* Bind Enter to custom handler that clears
-     * ghost text before accepting the line */
-    rl_bind_key('\r', cli_accept_line);
-    rl_bind_key('\n', cli_accept_line);
-}
-#else
-void CLI_configure_readline() {}
-void CLI_setup_hint_area(void) {}
-void CLI_cleanup_scroll_region(void) {}
-#endif
