+++ src/cli/CLIcore/CLIcore/CLIcore_script.c
@@ -497,912 +497,6 @@ int eval_cond_line(
  *       body3
  *   fi
  */
-void cli_exec_block_if(
-    char lines[][STRINGMAXLEN_CLICMDLINE],
-    int nlines
-)
-{
-    if(nlines < 2)
-    {
-        return;
-    }
-
-    /* Build a list of branches:
-     * Each branch has a condition line index
-     * and body range [start, end).
-     * The final else has cond_idx = -1. */
-
-    typedef struct
-    {
-        int cond_idx;
-        int body_start;
-        int body_end;
-    } Branch;
-
-    Branch branches[64];
-    int nbranch = 0;
-
-    /* First branch: the if line */
-    int body_s = 1;
-    /* Skip standalone "then" */
-    if(body_s < nlines)
-    {
-        const char *ts =
-            strip_ws(lines[body_s]);
-        if(strcmp(ts, "then") == 0)
-        {
-            body_s++;
-        }
-    }
-
-    branches[0].cond_idx = 0;
-    branches[0].body_start = body_s;
-    branches[0].body_end = nlines; /* Will be updated if elif/else found */
-    nbranch = 1;
-
-    /* Scan for elif/else at depth 0 */
-    int depth = 0;
-    for(int i = body_s; i < nlines; i++)
-    {
-        const char *ln = strip_ws(lines[i]);
-        if(starts_with(ln, "if ")
-           || starts_with(ln, "if\t"))
-        {
-            depth++;
-            continue;
-        }
-        if(strcmp(ln, "fi") == 0)
-        {
-            if(depth > 0)
-            {
-                depth--;
-                continue;
-            }
-            /* Should not happen since fi is not appended */
-            branches[nbranch - 1].body_end = i;
-            break;
-        }
-        if(depth > 0)
-        {
-            continue;
-        }
-        if(starts_with(ln, "elif ")
-           || starts_with(ln, "elif\t"))
-        {
-            branches[nbranch - 1].body_end
-                = i;
-            int bs = i + 1;
-            if(bs < nlines)
-            {
-                const char *t2 =
-                    strip_ws(lines[bs]);
-                if(strcmp(t2, "then") == 0)
-                {
-                    bs++;
-                }
-            }
-            if(nbranch < 64)
-            {
-                branches[nbranch].cond_idx
-                    = i;
-                branches[nbranch].body_start
-                    = bs;
-                branches[nbranch].body_end
-                    = nlines; /* Will be updated if else found */
-                nbranch++;
-            }
-        }
-        else if(strcmp(ln, "else") == 0)
-        {
-            branches[nbranch - 1].body_end
-                = i;
-            if(nbranch < 64)
-            {
-                branches[nbranch].cond_idx
-                    = -1; /* else */
-                branches[nbranch].body_start
-                    = i + 1;
-                branches[nbranch].body_end
-                    = nlines;
-                nbranch++;
-            }
-            break;
-        }
-    }
-
-    /* Evaluate branches in order */
-    for(int b = 0; b < nbranch; b++)
-    {
-        int run = 0;
-        if(branches[b].cond_idx < 0)
-        {
-            /* else — always true */
-            run = 1;
-        }
-        else
-        {
-            const char *cl2 = strip_ws(
-                lines[branches[b].cond_idx]);
-            int skip = 2; /* "if" */
-            if(starts_with(cl2, "elif"))
-            {
-                skip = 4;
-            }
-            run = eval_cond_line(
-                lines[branches[b].cond_idx],
-                skip);
-        }
-        if(run)
-        {
-            int bs = branches[b].body_start;
-            int be = branches[b].body_end;
-            if(be > bs)
-            {
-                cli_exec_lines(
-                    lines + bs, be - bs);
-            }
-            break;
-        }
-    }
-}
-
-
-/* ---- Parse while/do/done block ---- */
-
-/**
- * @brief Execute a while/do/done block
- *
- * Expected format:
- *   lines[0]: "while [ condition ]; do"
- *   ...body...
- *   "done"
- */
-void cli_exec_block_while(
-    char lines[][STRINGMAXLEN_CLICMDLINE],
-    int nlines
-)
-{
-    if(nlines < 2)
-    {
-        return;
-    }
-
-    /* Find body start (after "do") */
-    int body_start = 1;
-    int body_end = nlines;
-    int max_iter = 100000;
-
-    /* Skip standalone 'do' line from
-     * semicolon-split */
-    if(body_start < body_end)
-    {
-        const char *ds =
-            strip_ws(lines[body_start]);
-        if(strcmp(ds, "do") == 0)
-        {
-            body_start++;
-        }
-    }
-
-    for(int iter = 0; iter < max_iter; iter++)
-    {
-        /* Re-expand condition each iteration */
-        char condline[STRINGMAXLEN_CLICMDLINE];
-        strncpy(condline, lines[0],
-                STRINGMAXLEN_CLICMDLINE - 1);
-        condline[
-            STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-
-        /* Run expansion on condition */
-        cli_expand_fpsvar(
-            condline,
-            STRINGMAXLEN_CLICMDLINE);
-        cli_expand_env(
-            condline,
-            STRINGMAXLEN_CLICMDLINE);
-        cli_expand_arith(
-            condline,
-            STRINGMAXLEN_CLICMDLINE);
-
-        /* Parse condition */
-        const char *cl = strip_ws(condline);
-        cl += 5; /* skip "while" */
-        cl = strip_ws(cl);
-
-        int cond_result = 0;
-        if(*cl == '[')
-        {
-            cl++;
-            const char *end = strrchr(cl, ']');
-            if(end != NULL)
-            {
-                char cs[512];
-                int clen = (int)(end - cl);
-                if(clen >= (int) sizeof(cs))
-                {
-                    clen =
-                        (int) sizeof(cs) - 1;
-                }
-                memcpy(cs, cl, (size_t) clen);
-                cs[clen] = '\0';
-                cond_result =
-                    cli_eval_test(cs);
-            }
-        }
-        else
-        {
-            /* Check command exit status */
-            char ccmd[STRINGMAXLEN_CLICMDLINE];
-            strncpy(ccmd, cl, sizeof(ccmd) - 1);
-            ccmd[sizeof(ccmd) - 1] = '\0';
-            
-            char *semicolon = strstr(ccmd, ";");
-            if(semicolon)
-            {
-                char *do_ptr = strstr(semicolon, "do");
-                if(do_ptr)
-                {
-                    *semicolon = '\0';
-                }
-            }
-            
-            /* Strip trailing whitespace/semicolon */
-            int len = (int)strlen(ccmd);
-            while(len > 0 && (ccmd[len - 1] == ';' || ccmd[len - 1] == ' ' || ccmd[len - 1] == '\t'))
-            {
-                ccmd[--len] = '\0';
-            }
-            CLI_execute_string(ccmd);
-            cond_result = (cli_last_retval == 0) ? 1 : 0;
-        }
-
-        if(!cond_result)
-        {
-            break;
-        }
-
-        /* Execute body */
-        cli_continue_flag = 0;
-        cli_exec_lines(
-            lines + body_start,
-            body_end - body_start);
-
-        if(cli_break_flag)
-        {
-            cli_break_flag = 0;
-            break;
-        }
-    }
-}
-
-
-/* ---- Parse for/do/done block ---- */
-
-/**
- * @brief Execute select/do/done block
- *
- * Syntax:
- *   select VAR in v1 v2 ...; do
- *     body
- *   done
- */
-void cli_exec_block_select(
-    char lines[][STRINGMAXLEN_CLICMDLINE],
-    int  nlines
-)
-{
-    if(nlines < 2)
-    {
-        return;
-    }
-    /* Parse: select VAR in v1 v2 ... */
-    const char *hdr = strip_ws(lines[0]);
-    hdr += 7; /* skip 'select ' */
-    while(*hdr == ' '
-          || *hdr == '\t')
-    {
-        hdr++;
-    }
-    char vn[CLI_VAR_NAMELEN];
-    {
-        int vi = 0;
-        while(*hdr != '\0'
-              && *hdr != ' '
-              && *hdr != '\t'
-              && vi
-              < CLI_VAR_NAMELEN - 1)
-        {
-            vn[vi++] = *hdr++;
-        }
-        vn[vi] = '\0';
-    }
-    while(*hdr == ' '
-          || *hdr == '\t')
-    {
-        hdr++;
-    }
-    if(starts_with(hdr, "in "))
-    {
-        hdr += 3;
-    }
-    /* Collect values */
-    char sv[256][CLI_VAR_VALLEN];
-    int nsv = 0;
-    while(*hdr != '\0'
-          && nsv < 256)
-    {
-        while(*hdr == ' '
-              || *hdr == '\t')
-        {
-            hdr++;
-        }
-        if(*hdr == ';'
-           || *hdr == '\0')
-        {
-            break;
-        }
-        int vi = 0;
-        while(*hdr != '\0'
-              && *hdr != ' '
-              && *hdr != '\t'
-              && *hdr != ';'
-              && vi
-              < CLI_VAR_VALLEN - 1)
-        {
-            sv[nsv][vi++] = *hdr++;
-        }
-        sv[nsv][vi] = '\0';
-        nsv++;
-    }
-    if(nsv == 0)
-    {
-        return;
-    }
-    /* Loop: print menu, read, exec */
-    for(;;)
-    {
-        for(int i = 0;
-            i < nsv; i++)
-        {
-            printf("%d) %s\n",
-                   i + 1, sv[i]);
-        }
-        printf("#? ");
-        fflush(stdout);
-        char rb[64];
-        if(fgets(rb, sizeof(rb),
-                 stdin) == NULL)
-        {
-            break;
-        }
-        int ch =
-            (int) strtol(rb,
-                         NULL, 10);
-        if(ch >= 1 && ch <= nsv)
-        {
-            cli_var_set(
-                vn, sv[ch - 1]);
-        }
-        else
-        {
-            cli_var_set(vn, "");
-        }
-        cli_exec_lines(
-            lines + 1,
-            nlines - 1);
-    }
-}
-
-/**
- * @brief Execute a for/do/done block
- *
- * Expected format:
- *   lines[0]: "for VAR in val1 val2 ...; do"
- *   ...body...
- *   "done"
- */
-void cli_exec_block_for(
-    char lines[][STRINGMAXLEN_CLICMDLINE],
-    int nlines
-)
-{
-    if(nlines < 2)
-    {
-        return;
-    }
-
-    /* Check for arithmetic for:
-     * for ((init; cond; step)); do */
-    {
-        const char *af =
-            strip_ws(lines[0]);
-        af += 3; /* skip "for" */
-        af = strip_ws(af);
-        if(af[0] == '(' && af[1] == '(')
-        {
-            af += 2; /* skip "((" */
-            /* Find closing )) */
-            const char *ce =
-                strstr(af, "))");
-            if(ce != NULL)
-            {
-                char abuf[
-                    STRINGMAXLEN_CLICMDLINE
-                ];
-                int alen =
-                    (int)(ce - af);
-                memcpy(abuf, af,
-                       (size_t) alen);
-                abuf[alen] = '\0';
-                /* Split on ; */
-                char ainit[256] = "";
-                char acond[256] = "";
-                char astep[256] = "";
-                char *s1 =
-                    strchr(abuf, ';');
-                if(s1 != NULL)
-                {
-                    *s1 = '\0';
-                    strncpy(ainit,
-                            abuf, 255);
-                    char *s2 =
-                        strchr(
-                            s1 + 1,
-                            ';');
-                    if(s2 != NULL)
-                    {
-                        *s2 = '\0';
-                        strncpy(
-                            acond,
-                            s1 + 1,
-                            255);
-                        strncpy(
-                            astep,
-                            s2 + 1,
-                            255);
-                    }
-                }
-                /* Execute init */
-                {
-                    char einit[
-                        STRINGMAXLEN_CLICMDLINE
-                    ];
-                    snprintf(
-                        einit,
-                        sizeof(einit),
-                        "$(( %s ))",
-                        ainit);
-                    cli_expand_arith(
-                        einit,
-                        STRINGMAXLEN_CLICMDLINE
-                    );
-                }
-                /* Loop: eval cond,
-                 * exec body, eval step */
-                for(;;)
-                {
-                    char econd[
-                        STRINGMAXLEN_CLICMDLINE
-                    ];
-                    snprintf(
-                        econd,
-                        sizeof(econd),
-                        "$(( %s ))",
-                        acond);
-                    cli_expand_arith(
-                        econd,
-                        STRINGMAXLEN_CLICMDLINE
-                    );
-                    long cv =
-                        strtol(econd,
-                               NULL,
-                               10);
-                    if(cv == 0)
-                    {
-                        break;
-                    }
-                    cli_exec_lines(
-                        lines + 1,
-                        nlines - 1);
-                    /* step */
-                    {
-                        char estep[
-                            STRINGMAXLEN_CLICMDLINE
-                        ];
-                        snprintf(
-                            estep,
-                            sizeof(estep),
-                            "$(( %s ))",
-                            astep);
-                        cli_expand_arith(
-                            estep,
-                            STRINGMAXLEN_CLICMDLINE
-                        );
-                    }
-                }
-                return;
-            }
-        }
-    }
-
-    /* Parse: for VAR in val1 val2 ... */
-    const char *fl = strip_ws(lines[0]);
-    fl += 3; /* skip "for" */
-    fl = strip_ws(fl);
-
-    /* Get variable name */
-    char varname[CLI_VAR_NAMELEN];
-    {
-        int vn = 0;
-        while(*fl != '\0' && *fl != ' '
-              && *fl != '\t'
-              && vn < CLI_VAR_NAMELEN - 1)
-        {
-            varname[vn++] = *fl++;
-        }
-        varname[vn] = '\0';
-    }
-    fl = strip_ws(fl);
-
-    /* Skip "in" */
-    if(strncmp(fl, "in ", 3) == 0
-       || strncmp(fl, "in\t", 3) == 0)
-    {
-        fl += 2;
-        fl = strip_ws(fl);
-    }
-
-    /* Collect values (strip trailing ;do) */
-    char vallist[STRINGMAXLEN_CLICMDLINE];
-    strncpy(vallist, fl,
-            STRINGMAXLEN_CLICMDLINE - 1);
-    vallist[
-        STRINGMAXLEN_CLICMDLINE - 1] = '\0';
-
-    /* Remove trailing "; do" or ";do" */
-    {
-        char *semi = strstr(vallist, ";");
-        if(semi != NULL)
-        {
-            *semi = '\0';
-        }
-    }
-
-    /* Strip trailing whitespace */
-    {
-        size_t vl = strlen(vallist);
-        while(vl > 0
-              && (vallist[vl - 1] == ' '
-                  || vallist[vl - 1] == '\t'
-                  || vallist[vl - 1] == '\n'))
-        {
-            vallist[--vl] = '\0';
-        }
-    }
-
-    int body_start = 1;
-    int body_end = nlines;
-
-    /* Skip standalone 'do' line */
-    if(body_start < body_end)
-    {
-        const char *ds =
-            strip_ws(lines[body_start]);
-        if(strcmp(ds, "do") == 0)
-        {
-            body_start++;
-        }
-    }
-
-    /* Iterate over values */
-    char *saveptr = NULL;
-    char *val = strtok_r(vallist, " \t",
-                         &saveptr);
-    while(val != NULL)
-    {
-        cli_var_set(varname, val);
-
-        cli_continue_flag = 0;
-        cli_exec_lines(
-            lines + body_start,
-            body_end - body_start);
-
-        if(cli_break_flag)
-        {
-            cli_break_flag = 0;
-            break;
-        }
-
-        val = strtok_r(NULL, " \t", &saveptr);
-    }
-}
-
-
-/* ============================================================
- *  User-Defined Functions
- * ============================================================
- */
-
-CLI_FUNC cli_funcs[CLI_MAX_FUNCS];
-
-/**
- * @brief Register a new user function
- */
-void cli_func_define(
-    const char *name,
-    char body[][STRINGMAXLEN_CLICMDLINE],
-    int nbody
-)
-{
-    /* Update existing */
-    for(int i = 0; i < CLI_MAX_FUNCS; i++)
-    {
-        if(cli_funcs[i].used
-           && strcmp(cli_funcs[i].name, name)
-              == 0)
-        {
-            cli_funcs[i].nbody = nbody;
-            for(int j = 0; j < nbody; j++)
-            {
-                strncpy(
-                    cli_funcs[i].body[j],
-                    body[j],
-                    STRINGMAXLEN_CLICMDLINE - 1
-                );
-            }
-            return;
-        }
-    }
-    /* Find empty slot */
-    for(int i = 0; i < CLI_MAX_FUNCS; i++)
-    {
-        if(!cli_funcs[i].used)
-        {
-            strncpy(cli_funcs[i].name, name,
-                    CLI_FUNC_NAMELEN - 1);
-            cli_funcs[i].name[
-                CLI_FUNC_NAMELEN - 1] = '\0';
-            cli_funcs[i].nbody = nbody;
-            cli_funcs[i].used = 1;
-            for(int j = 0; j < nbody; j++)
-            {
-                strncpy(
-                    cli_funcs[i].body[j],
-                    body[j],
-                    STRINGMAXLEN_CLICMDLINE - 1
-                );
-            }
-            return;
-        }
-    }
-    printf("Error: function table full "
-           "(max %d)\n", CLI_MAX_FUNCS);
-}
-
-
-/* ============================================================
- *  Block Intercept — main entry point
- * ============================================================
- *
- * Called from CLI_execute_line() before any other
- * processing. Returns 1 if the line was consumed
- * (buffered or block completed).
- */
-
-/**
- * @brief Intercept line for flow control
- *
- * @param line  The raw command line
- * @return 1 if consumed, 0 if not
- */
-/* ============================================================
- *  Case/esac Evaluator
- * ============================================================
- *
- * Syntax:
- *   case <word> in
- *     pattern1) cmd1 ;;
- *     pat2|pat3) cmd2 ;;
- *     *) default ;;
- *   esac
- */
-void cli_exec_block_case(
-    char (*lines)[STRINGMAXLEN_CLICMDLINE],
-    int    nlines
-)
-{
-    /* Line 0 = "case <word> in" */
-    const char *hdr = strip_ws(lines[0]);
-    hdr += 4; /* skip "case" */
-    hdr = strip_ws(hdr);
-    char word[256];
-    {
-        int wi = 0;
-        while(*hdr != '\0'
-              && *hdr != ' '
-              && *hdr != '\t'
-              && wi < 255)
-        {
-            word[wi++] = *hdr++;
-        }
-        word[wi] = '\0';
-    }
-    /* Expand word */
-    cli_expand_env(word, 256);
-
-    /* Scan patterns: "pat) body ;;" */
-    for(int i = 1; i < nlines; i++)
-    {
-        const char *lp =
-            strip_ws(lines[i]);
-        /* Find closing ')' */
-        const char *cp = strchr(lp, ')');
-        if(cp == NULL)
-        {
-            continue;
-        }
-        /* Extract pattern(s) */
-        char pat[256];
-        int plen = (int)(cp - lp);
-        if(plen >= 256)
-        {
-            plen = 255;
-        }
-        memcpy(pat, lp, (size_t) plen);
-        pat[plen] = '\0';
-
-        /* Check match (supports pat1|pat2
-         * and * wildcard) */
-        int matched = 0;
-        {
-            char ptmp[256];
-            strncpy(ptmp, pat,
-                    sizeof(ptmp) - 1);
-            ptmp[sizeof(ptmp) - 1] = '\0';
-            char *psave = NULL;
-            char *pp =
-                strtok_r(ptmp, "|",
-                         &psave);
-            while(pp != NULL)
-            {
-                /* strip ws */
-                while(*pp == ' '
-                      || *pp == '\t')
-                {
-                    pp++;
-                }
-                if(strcmp(pp, "*") == 0
-                   || strcmp(pp, word)
-                   == 0)
-                {
-                    matched = 1;
-                    break;
-                }
-                pp = strtok_r(NULL,
-                              "|",
-                              &psave);
-            }
-        }
-        if(!matched)
-        {
-            continue;
-        }
-
-        /* Collect body lines until ;; */
-        const char *body_start = cp + 1;
-        while(*body_start == ' '
-              || *body_start == '\t')
-        {
-            body_start++;
-        }
-        /* If body is on same line */
-        if(*body_start != '\0')
-        {
-            /* Strip ;; from end */
-            char cmdline[STRINGMAXLEN_CLICMDLINE];
-            strncpy(cmdline, body_start,
-                    STRINGMAXLEN_CLICMDLINE
-                    - 1);
-            cmdline[STRINGMAXLEN_CLICMDLINE
-                    - 1] = '\0';
-            {
-                int cl =
-                    (int) strlen(cmdline);
-                while(cl > 1
-                      && cmdline[cl - 1]
-                      == ';'
-                      && cmdline[cl - 2]
-                      == ';')
-                {
-                    cmdline[cl - 2] = '\0';
-                    cl -= 2;
-                }
-                /* Trim trailing ws */
-                while(cl > 0
-                      && (cmdline[cl - 1]
-                          == ' '
-                          || cmdline[cl - 1]
-                          == '\t'))
-                {
-                    cmdline[--cl] = '\0';
-                }
-            }
-            if(strlen(cmdline) > 0)
-            {
-                strncpy(
-                    data.CLIcmdline,
-                    cmdline,
-                    STRINGMAXLEN_CLICMDLINE
-                    - 1);
-                CLI_execute_line();
-            }
-        }
-        else
-        {
-            /* Multi-line body */
-            for(int j = i + 1;
-                j < nlines; j++)
-            {
-                const char *bl =
-                    strip_ws(lines[j]);
-                if(strcmp(bl, ";;") == 0)
-                {
-                    break;
-                }
-                /* Strip trailing ;; */
-                char cmd2[
-                    STRINGMAXLEN_CLICMDLINE];
-                strncpy(cmd2, bl,
-                        STRINGMAXLEN_CLICMDLINE
-                        - 1);
-                cmd2[
-                    STRINGMAXLEN_CLICMDLINE
-                    - 1] = '\0';
-                {
-                    int c2l =
-                        (int) strlen(cmd2);
-                    int ends_dsemi = 0;
-                    while(c2l > 1
-                          && cmd2[c2l - 1]
-                          == ';'
-                          && cmd2[c2l - 2]
-                          == ';')
-                    {
-                        cmd2[c2l - 2] =
-                            '\0';
-                        c2l -= 2;
-                        ends_dsemi = 1;
-                    }
-                    while(c2l > 0
-                          && (cmd2[c2l - 1]
-                              == ' '
-                              || cmd2[
-                                  c2l - 1]
-                              == '\t'))
-                    {
-                        cmd2[--c2l] = '\0';
-                    }
-                    if(strlen(cmd2) > 0)
-                    {
-                        strncpy(
-                            data.CLIcmdline,
-                            cmd2,
-                            STRINGMAXLEN_CLICMDLINE
-                            - 1);
-                        CLI_execute_line();
-                    }
-                    if(ends_dsemi)
-                    {
-                        break;
-                    }
-                }
-            }
-        }
-        return; /* first match only */
-    }
-}
-
-
 int cli_script_intercept(const char *line)
 {
     const char *p = strip_ws(line);
